common: rename utils.h -> bi_utils.h and utils.cpp -> bi_utils.cpp
