+++ src/libs/battle_interface/include/bi_utils.h
