+++ src/libs/battle_interface/src/bi_utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "bi_utils.h"
 
 #include "core.h"
 #include "string_compare.hpp"
