+++ src/libs/battle_interface/src/item_entity/item_entity.cpp
@@ -1,5 +1,5 @@
 #include "item_entity.h"
-#include "utils.h"
+#include "bi_utils.h"
 #include "entity.h"
 #include "core.h"
 #include "message.h"
