+++ src/libs/battle_interface/src/log_and_action.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "utils.h"
+#include "bi_utils.h"
 
 class ILogAndActions : public Entity
 {
